requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
Django==2.2
djangorestframework==3.9.2
kombu==3.0.37
pika==1.0.1
psycopg2==2.8.2
pytz==2019.1
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.1
vine==1.3.0
